removed terrible gpt data handling
got rid of gpt suggested data handling and implemented my own. encoded white and black as part of the input tensor, not sure how i will be able to make use of it just yet.

thinking of doing some sort of LSTM model that generates an encoded move based on the board state as well as the current sequence in the game. hopefully can eventually teach themed tactics like the greek gift and some trappy lines