letsfxckinGOOOOOOOOOOOOOOOOOO
ladies and gentlemen we have liftoff.

model training is currently around 85% accuracy.

biggest problem currently is getting out of stages of the game (opening, middlegame) and recommending illegal moves over and over

need to somehow intervene in this scenario, and force the model to come up with a legal move. somehow limiting evaluation to only legal moves would be interesting, and im pretty sure something like that is in my gpt logs